+++ iExtra/Extensions/Url/Url+LocalFile.swift
@@ -1,0 +1,20 @@
+//
+//  Url+LocalFile.swift
+//  iExtra
+//
+//  Created by Daniel Saidi on 2018-08-06.
+//  Copyright © 2018 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension URL {
+    
+    public var localFileExists: Bool {
+        do {
+            return try checkPromisedItemIsReachable()
+        } catch {
+            return false
+        }
+    }
+}